+++ include/Dice/hypertrie/internal/ConfigHypertrieDepthLimit.hpp
