+++ include/Dice/hypertrie/internal/raw/iterator/Iterator.hpp
@@ -1,8 +1,8 @@
 #ifndef HYPERTRIE_RAWITERATOR_HPP
 #define HYPERTRIE_RAWITERATOR_HPP
 
-#include "Dice/hypertrie/internal/node_based/raw/storage/NodeContext.hpp"
-#include "Dice/hypertrie/internal/node_based/ConfigHypertrieDepthLimit.hpp"
+#include "Dice/hypertrie/internal/ConfigHypertrieDepthLimit.hpp"
+#include "Dice/hypertrie/internal/raw/storage/NodeContext.hpp"
 #include "Dice/hypertrie/internal/util/IntegralTemplatedTuple.hpp"
 namespace hypertrie::internal::node_based::raw {
