+++ include/Dice/hypertrie/internal/raw/storage/Entry.hpp
@@ -1,7 +1,7 @@
 #ifndef HYPERTRIE_ENTRY_HPP
 #define HYPERTRIE_ENTRY_HPP
 
-#include "Dice/hypertrie/internal/node_based/raw/Hypertrie_internal_traits.hpp"
+#include "Dice/hypertrie/internal/raw/Hypertrie_internal_traits.hpp"
 
 
 namespace hypertrie::internal::node_based::raw {
