+++ include/Dice/hypertrie/internal/raw/storage/NodeModificationPlan.hpp
@@ -1,9 +1,9 @@
 #ifndef HYPERTRIE_NODEMODIFICATIONPLAN_HPP
 #define HYPERTRIE_NODEMODIFICATIONPLAN_HPP
 
-#include "Dice/hypertrie/internal/node_based/raw/node/TensorHash.hpp"
-#include "Dice/hypertrie/internal/node_based/raw/storage/Entry.hpp"
-#include "Dice/hypertrie/internal/node_based/raw/Hypertrie_internal_traits.hpp"
+#include "Dice/hypertrie/internal/raw/Hypertrie_internal_traits.hpp"
+#include "Dice/hypertrie/internal/raw/node/TensorHash.hpp"
+#include "Dice/hypertrie/internal/raw/storage/Entry.hpp"
 
 
 namespace hypertrie::internal::node_based::raw {
