+++ tests/node_based/TestNodeContextRandomized.hpp
@@ -3,7 +3,7 @@
 
 #include <iterator>
 
-#include <Dice/hypertrie/internal/node_based/raw/storage/NodeContext.hpp>
+#include <Dice/hypertrie/internal/raw/storage/NodeContext.hpp>
 
 #include <fmt/format.h>
