+++ tests/node_based/TestRawDiagonal.hpp
@@ -1,15 +1,15 @@
 #ifndef HYPERTRIE_TESTRAWDIAGONAL_HPP
 #define HYPERTRIE_TESTRAWDIAGONAL_HPP
 
-#include <Dice/hypertrie/internal/node_based/raw/storage/NodeContext.hpp>
-#include <Dice/hypertrie/internal/node_based/interface/Hypertrie_traits.hpp>
-#include <Dice/hypertrie/internal/node_based/raw/iterator/Diagonal.hpp>
+#include <Dice/hypertrie/internal/raw/storage/NodeContext.hpp>
+#include <Dice/hypertrie/internal/interface/Hypertrie_traits.hpp>
+#include <Dice/hypertrie/internal/raw/iterator/Diagonal.hpp>
 #include <fmt/format.h>
 
 #include "../utils/AssetGenerator.hpp"
 #include "../utils/NameOfType.hpp"
 #include "TestTensor.hpp"
-namespace hypertrie::tests::node_based::raw::node_context::iterator_test {
+namespace hypertrie::tests::node_based::raw::node_context::diagonal_test {
 
 	using namespace hypertrie::tests::utils;
 
@@ -18,7 +18,7 @@ namespace hypertrie::tests::node_based::raw::node_context::iterator_test {
 	using namespace hypertrie::internal::node_based;
 
 
-	TEST_CASE("bool depth 1 uncompressed", "[raw diagonal]") {
+	TEST_CASE("bool depth 1 uncompressed d", "[raw diagonal]") {
 		using tri = default_bool_Hypertrie_internal_t;
 		constexpr const size_t depth = 1;
 		using key_part_type = typename tri::key_part_type;
@@ -45,7 +45,7 @@ namespace hypertrie::tests::node_based::raw::node_context::iterator_test {
 		}
 	}
 
-	TEST_CASE("bool depth 2 uncompressed", "[raw diagonal]") {
+	TEST_CASE("bool depth 2 uncompressed d", "[raw diagonal]") {
 		using tri = default_bool_Hypertrie_internal_t;
 		constexpr const size_t depth = 2;
 		using key_part_type = typename tri::key_part_type;
@@ -79,7 +79,7 @@ namespace hypertrie::tests::node_based::raw::node_context::iterator_test {
 		}
 	}
 
-	TEST_CASE("bool depth 2 compressed", "[raw diagonal]") {
+	TEST_CASE("bool depth 2 compressed d", "[raw diagonal]") {
 		using tri = default_bool_Hypertrie_internal_t;
 		constexpr const size_t depth = 2;
 		using key_part_type = typename tri::key_part_type;
